Remove cyclical map importer reference
